move StatementLinesCollector to utils
